feat: create BaseChart component for reusable chart functionality
- Added Base.vue as a reusable chart component with ECharts integration.
- Refactored Followers.vue, Subscribers.vue, Viewers.vue, and Messages.vue to utilize BaseChart for rendering charts.
- Updated chart options and data handling in the respective components.
- Enhanced styling and responsiveness of the chart components.

feat: implement lazy loading for chart components in dashboard

- Replaced direct chart component usage with lazy-loaded versions in the dashboard index.vue.
- Added LazyChartMessages component for displaying chat messages count.

fix(socket): enhance WebSocket data handling

- Added chatMessages and topChatters to the stream data handling in the socket.io client plugin.
- Updated stream history structure in the streams store to accommodate new data fields.
