+++ components/Chart/Messages.vue
@@ -1,12 +1,12 @@
 <template>
   <div class="chart-container">
     <BaseChart
-      :title="'Obserwujący na żywo'"
-      :legend="'Obserwujący'"
-      :yAxisName="'Obserwujący'"
-      :color="'#FFD700'"
+      :title="'Wiadomości na czacie w czasie'"
+      :legend="'Wiadomości'"
+      :yAxisName="'Wiadomości'"
+      :color="'#C800FF'"
       :areaColor="areaColor"
-      :data="followerData"
+      :data="messagesData"
       :streamExists="!!stream"
     />
   </div>
@@ -21,17 +21,17 @@ const streamsStore = useStreamsStore()
 
 const stream = computed(() => streamsStore.getStreamByStreamerName(props.streamerName))
 
-const followerData = computed(() => {
+const messagesData = computed((): [Date, number][] => {
   const history = streamsStore.getHistoryByStreamerName(props.streamerName)
-  return history?.followers?.map((v: any) => [new Date(v.timestamp), v.count]) || [];
+  return history?.chatMessages?.map((v: any) => [new Date(v.timestamp), v.count]) || [];
 })
 
-const areaColor = {
+const areaColor: any = {
   type: 'linear',
   x: 0, y: 0, x2: 0, y2: 1,
   colorStops: [
-    { offset: 0, color: 'rgba(255, 215, 0, 0.7)' },
-    { offset: 1, color: 'rgba(255, 215, 0, 0.1)' }
+    { offset: 0, color: 'rgba(200, 0, 255, 0.7)' },
+    { offset: 1, color: 'rgba(200, 0, 255, 0.1)' }
   ]
 }
 </script>
