+++ package.json
@@ -19,11 +19,11 @@
   "require": "./dist/modern/umd/dash.all.min.js",
   "dependencies": {
     "@mdi/font": "^7.4.47",
-    "@nuxt/icon": "^1.12.0",
+    "@nuxt/icon": "^1.13.0",
     "@nuxt/image": "^1.10.0",
     "@nuxt/kit": "^3.17.3",
     "@nuxtjs/color-mode": "3.5.2",
-    "@nuxtjs/i18n": "9.5.3",
+    "@nuxtjs/i18n": "9.5.4",
     "@pinia/nuxt": "^0.11.0",
     "@types/crypto-js": "^4.2.2",
     "@vueuse/core": "^13.2.0",
@@ -35,18 +35,18 @@
     "http-status-codes": "^2.3.0",
     "lodash-es": "^4.17.21",
     "nuxt": "^3.17.3",
-    "nuxt-echarts": "0.3.0",
+    "nuxt-echarts": "0.3.1",
     "pinia-plugin-persistedstate": "^4.3.0",
     "socket.io-client": "^4.8.1",
-    "vue": "latest",
-    "vue-router": "latest",
+    "vue": "^3.5.14",
+    "vue-router": "^4.5.1",
     "vuetify-nuxt-module": "0.18.6"
   },
   "devDependencies": {
     "@types/lodash-es": "^4.17.12",
     "axios": "^1.9.0",
     "vite-plugin-vuetify": "^2.1.1",
-    "vitest": "^3.1.3",
+    "vitest": "^3.1.4",
     "vuetify": "^3.8.5"
   }
 }
